pyproject.toml
[tool.poetry]
name = "users-crud-drf"
version = "0.1.0"
description = ""
authors = ["electr0n4ik <andreyshka3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
